requirements.txt
locust==2.16.1
locust-plugins==4.0.0
paho-mqtt==1.6.1
websocket-client==1.6.2
ujson
